LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-vertical-lr-002-expected.txt


FAIL .grid 1 assert_equals: 
<div class="grid directionRTL">
  <div class="paddingBottom50Percent" data-expected-padding-bottom="50" data-expected-width="10" data-expected-height="60">X</div>
  <div data-offset-x="10" data-offset-y="400" data-expected-width="10" data-expected-height="100"></div>
</div>
height expected 60 but got 50
FAIL .grid 2 assert_equals: 
<div class="grid directionRTL">
  <div class="paddingTop50Percent" data-expected-padding-top="50" data-expected-width="10" data-expected-height="60">X</div>
  <div data-offset-x="10" data-offset-y="400" data-expected-width="10" data-expected-height="100"></div>
</div>
height expected 60 but got 50
PASS .grid 3 
PASS .grid 4 
FAIL .grid 5 assert_equals: 
<div class="grid">
  <div class="paddingBottom50Percent" data-expected-padding-bottom="50" data-expected-width="10" data-expected-height="60">X</div>
  <div data-offset-x="10" data-offset-y="0" data-expected-width="10" data-expected-height="100"></div>
</div>
height expected 60 but got 50
FAIL .grid 6 assert_equals: 
<div class="grid">
  <div class="paddingTop50Percent" data-expected-padding-top="50" data-expected-width="10" data-expected-height="60">X</div>
  <div data-offset-x="10" data-offset-y="0" data-expected-width="10" data-expected-height="100"></div>
</div>
height expected 60 but got 50
PASS .grid 7 
PASS .grid 8 
Direction LTR

Item padding-left: 50%;
X
Item padding-right: 50%;
X
Item padding-top: 50%;
X
Item padding-bottom: 50%;
X
Direction RTL

Item padding-left: 50%;
X
Item padding-right: 50%;
X
Item padding-top: 50%;
X
Item padding-bottom: 50%;
X
